update stdsc dev-v2.0.0 and partial fix

+++ opsica/opsica_cloud/opsica_cloud_state.cpp
@@ -54,13 +54,13 @@ struct StateInit::Impl
         }
     }
 
-    uint64_t id(void) const
+    int32_t id(void) const
     {
         return id_;
     }
 
 private:
-    uint64_t id_;
+    int32_t id_;
     std::size_t count_;
     std::mutex mutex_;
 };
@@ -104,13 +104,13 @@ struct StateConnected::Impl
         }
     }
 
-    uint64_t id(void) const
+    int32_t id(void) const
     {
         return id_;
     }
 
 private:
-    uint64_t id_;
+    int32_t id_;
     bool is_stored_pubkey_;
     bool is_stored_dataA_;
     bool is_stored_dataB_;
@@ -140,13 +140,13 @@ struct StateReady::Impl
         }
     }
 
-    uint64_t id(void) const
+    int32_t id(void) const
     {
         return id_;
     }
 
 private:
-    uint64_t id_;
+    int32_t id_;
     std::mutex mutex_;
 };
 
@@ -182,13 +182,13 @@ struct StateComputed::Impl
         }
     }
 
-    uint64_t id(void) const
+    int32_t id(void) const
     {
         return id_;
     }
 
 private:
-    uint64_t id_;
+    int32_t id_;
     std::mutex mutex_;
 };
 
@@ -202,13 +202,13 @@ struct StateExit::Impl
     {
     }
 
-    uint64_t id(void) const
+    int32_t id(void) const
     {
         return id_;
     }
 
 private:
-    uint64_t id_;
+    int32_t id_;
     std::mutex mutex_;
 };
 
@@ -229,7 +229,7 @@ void StateInit::set(stdsc::StateContext& sc, uint64_t event)
     pimpl_->set(sc, event);
 }
 
-uint64_t StateInit::id(void) const
+int32_t StateInit::id(void) const
 {
     return pimpl_->id();
 }
@@ -256,7 +256,7 @@ void StateConnected::set(stdsc::StateContext& sc, uint64_t event)
     pimpl_->set(sc, event);
 }
 
-uint64_t StateConnected::id(void) const
+int32_t StateConnected::id(void) const
 {
     return pimpl_->id();
 }
@@ -278,7 +278,7 @@ void StateReady::set(stdsc::StateContext& sc, uint64_t event)
     pimpl_->set(sc, event);
 }
 
-uint64_t StateReady::id(void) const
+int32_t StateReady::id(void) const
 {
     return pimpl_->id();
 }
@@ -300,7 +300,7 @@ void StateComputed::set(stdsc::StateContext& sc, uint64_t event)
     pimpl_->set(sc, event);
 }
 
-uint64_t StateComputed::id(void) const
+int32_t StateComputed::id(void) const
 {
     return pimpl_->id();
 }
@@ -322,7 +322,7 @@ void StateExit::set(stdsc::StateContext& sc, uint64_t event)
     pimpl_->set(sc, event);
 }
 
-uint64_t StateExit::id(void) const
+int32_t StateExit::id(void) const
 {
     return pimpl_->id();
 }
