+++ opsica/opsica_cloud/opsica_cloud_state.hpp
@@ -28,7 +28,7 @@ namespace opsica_cloud
 /**
  * @brief Enumeration for state of cloud.
  */
-enum StateId_t : uint64_t
+enum StateId_t : int32_t
 {
     kStateNil = 0,
     kStateInit = 1,
@@ -61,7 +61,7 @@ struct StateInit : public stdsc::State
     static std::shared_ptr<State> create(size_t initial_connection_count = 0);
     StateInit(size_t initial_connection_count = 0);
     virtual void set(stdsc::StateContext& sc, uint64_t event) override;
-    virtual uint64_t id(void) const override;
+    virtual int32_t id(void) const override;
 
 private:
     struct Impl;
@@ -80,7 +80,7 @@ struct StateConnected : public stdsc::State
     StateConnected(bool is_stored_pubkey = false, bool is_stored_dataA = false,
                    bool is_stored_dataB = false);
     virtual void set(stdsc::StateContext& sc, uint64_t event) override;
-    virtual uint64_t id(void) const override;
+    virtual int32_t id(void) const override;
 
 private:
     struct Impl;
@@ -95,7 +95,7 @@ struct StateReady : public stdsc::State
     static std::shared_ptr<State> create();
     StateReady(void);
     virtual void set(stdsc::StateContext& sc, uint64_t event) override;
-    virtual uint64_t id(void) const override;
+    virtual int32_t id(void) const override;
 
 private:
     struct Impl;
@@ -110,7 +110,7 @@ struct StateComputed : public stdsc::State
     static std::shared_ptr<State> create();
     StateComputed(void);
     virtual void set(stdsc::StateContext& sc, uint64_t event) override;
-    virtual uint64_t id(void) const override;
+    virtual int32_t id(void) const override;
 
 private:
     struct Impl;
@@ -125,7 +125,7 @@ struct StateExit : public stdsc::State
     static std::shared_ptr<State> create();
     StateExit(void);
     virtual void set(stdsc::StateContext& sc, uint64_t event) override;
-    virtual uint64_t id(void) const override;
+    virtual int32_t id(void) const override;
 
 private:
     struct Impl;
